# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
  dao.Company:
    properties:
      inn:
        example: "7707083893"
        type: string
      name:
        example: ООО 'Парк'
        maxLength: 150
        minLength: 2
        type: string
      website:
        example: https://www.rusprofile.ru
        type: string
    required:
    - inn
    type: object
  dao.Industry:
    properties:
      avgSalary:
        type: number
      avgSalaryCad:
        type: number
      avgWorkersNum:
        type: number
      avgWorkersNumCad:
        type: number
    required:
    - avgSalary
    - avgSalaryCad
    - avgWorkersNum
    - avgWorkersNumCad
    type: object
  dao.Me:
    properties:
      biography:
        example: I'd like to relax
        type: string
      city:
        example: Москва
        type: string
      country:
        example: Россия
        type: string
      email:
        example: myemail@gmail.com
        type: string
      fatherName:
        example: Ivanovich
        type: string
      firstName:
        example: Ivan
        type: string
      lastName:
        example: Ivanov
        type: string
      name:
        example: user94
        type: string
      position:
        example: Director
        type: string
      role:
        example: USER
        type: string
    required:
    - email
    - firstName
    - lastName
    - name
    - role
    type: object
  dto.Email:
    properties:
      email:
        example: myemail@gmail.com
        type: string
    required:
    - email
    type: object
  dto.History:
    properties:
      accountingServices:
        type: boolean
      companyName:
        maxLength: 150
        minLength: 2
        type: string
      constructionFacilitiesArea:
        maximum: 0
        type: integer
      districtTitle:
        type: string
      equipmentType:
        type: string
      facilityType:
        type: string
      fullTimeEmployees:
        maximum: 0
        type: integer
      industryBranch:
        type: string
      landArea:
        maximum: 0
        type: integer
      other:
        type: string
      patent:
        type: boolean
    required:
    - accountingServices
    - companyName
    - constructionFacilitiesArea
    - districtTitle
    - equipmentType
    - facilityType
    - fullTimeEmployees
    - industryBranch
    - landArea
    - other
    - patent
    type: object
  dto.IndustryType:
    properties:
      type:
        maxLength: 150
        type: string
    required:
    - type
    type: object
  dto.SignIn:
    properties:
      email:
        example: myemail@gmail.com
        type: string
      password:
        example: bob126
        maxLength: 20
        minLength: 4
        type: string
    required:
    - email
    - password
    type: object
  dto.SignUp:
    properties:
      biography:
        example: I'd like to relax
        maxLength: 1024
        type: string
      city:
        example: Москва
        type: string
      company:
        $ref: '#/definitions/dao.Company'
      country:
        example: Россия
        type: string
      email:
        example: myemail@gmail.com
        type: string
      fatherName:
        example: Ivanovich
        maxLength: 30
        minLength: 2
        type: string
      firstName:
        example: Ivan
        maxLength: 30
        minLength: 2
        type: string
      lastName:
        example: Ivanov
        maxLength: 30
        minLength: 2
        type: string
      password:
        example: bob126
        maxLength: 20
        minLength: 4
        type: string
      position:
        example: Director
        maxLength: 50
        minLength: 2
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    type: object
  dto.UpdateCompany:
    properties:
      inn:
        example: "7707083893"
        type: string
      name:
        example: ООО 'Парк'
        maxLength: 150
        minLength: 2
        type: string
      website:
        example: https://www.rusprofile.ru
        type: string
    type: object
  dto.UpdateEmail:
    properties:
      newEmail:
        example: myemail@gmail.com
        type: string
      password:
        example: mob126
        maxLength: 20
        minLength: 4
        type: string
    required:
    - newEmail
    - password
    type: object
  dto.UpdatePassword:
    properties:
      code:
        example: N1OSP
        type: string
      email:
        example: myemail@gmail.com
        type: string
      newPassword:
        example: mob126
        maxLength: 20
        minLength: 4
        type: string
    required:
    - code
    - email
    - newPassword
    type: object
  dto.UpdateUser:
    properties:
      biography:
        example: I'd like to relax
        maxLength: 1024
        type: string
      city:
        example: Москва
        type: string
      country:
        example: Россия
        type: string
      fatherName:
        example: Ivanovich
        maxLength: 30
        minLength: 2
        type: string
      firstName:
        example: Ivan
        maxLength: 30
        minLength: 2
        type: string
      lastName:
        example: Ivanov
        maxLength: 30
        minLength: 2
        type: string
      position:
        example: Director
        maxLength: 50
        minLength: 2
        type: string
    type: object
  errs.MyError:
    description: All native errors must be this type
    properties:
      advice:
        example: Try to send request later
        type: string
      message:
        example: Exception was occurred
        type: string
    type: object
info:
  contact:
    email: matvey-sizov@mail.ru
    name: Contact us
    url: https://github.com/while-act/hackathon-backend/issues/new/choose
  description: It's an API interacting with While.act using Golang
  title: While.act API
  version: "1.0"
paths:
  /auth/session:
    delete:
      description: Get cookie and delete them from db
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      summary: Delete cookie session
      tags:
      - Session
    get:
      description: Returns detail information about me
      responses:
        "200":
          description: Info about session
          schema:
            $ref: '#/definitions/dao.Me'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errs.MyError'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      security:
      - ApiKeyAuth: []
      summary: Get detail data about the user by session
      tags:
      - Session
  /auth/sign-in:
    post:
      description: Compare the user's password with an existing user's password. If
        it matches, create session of this user
      parameters:
      - description: User's email, password
        in: body
        name: SignIn
        required: true
        schema:
          $ref: '#/definitions/dto.SignIn'
      responses:
        "200":
          description: user's session
          schema:
            type: string
        "400":
          description: Data is not valid
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      summary: Sign in by password
      tags:
      - Auth
  /auth/sign-up:
    post:
      description: Compare the user's password with an existing user's password. If
        it matches, create session of the user. If the user does not exist, create
        new user
      parameters:
      - description: User's email, password, firstName, lastName, inn
        in: body
        name: SignUp
        required: true
        schema:
          $ref: '#/definitions/dto.SignUp'
      responses:
        "200":
          description: user's session
          schema:
            type: string
        "400":
          description: Data is not valid
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      summary: Sign up by password
      tags:
      - Auth
  /calc:
    get:
      description: Returns PDF file, gotten from body
      parameters:
      - description: Completed application form
        in: body
        name: from
        required: true
        schema:
          $ref: '#/definitions/dto.History'
      responses:
        "200":
          description: PDF file
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      summary: Generate PDF from body
      tags:
      - Calc
    post:
      description: Saves given values to user's history
      parameters:
      - description: Completed application form
        in: body
        name: from
        required: true
        schema:
          $ref: '#/definitions/dto.History'
      responses:
        "200":
          description: OK
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      security:
      - ApiKeyAuth: []
      summary: Save calc data to history
      tags:
      - Calc
  /calc/industry:
    get:
      description: Returns detail information about industry
      parameters:
      - description: Industry Type
        in: body
        name: industry_name
        required: true
        schema:
          $ref: '#/definitions/dto.IndustryType'
      responses:
        "200":
          description: Info about industry
          schema:
            $ref: '#/definitions/dao.Industry'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: Industry doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      summary: Get data about industry
      tags:
      - Calc
  /company:
    get:
      description: Returns information about company by session
      responses:
        "200":
          description: Info about company
          schema:
            $ref: '#/definitions/dao.Company'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: Company doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      security:
      - ApiKeyAuth: []
      summary: Get data about company by session
      tags:
      - Company
    patch:
      description: Updates information about company by INN
      parameters:
      - description: Company
        in: body
        name: updCompany
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCompany'
      responses:
        "200":
          description: OK
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errs.MyError'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: Something doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      security:
      - ApiKeyAuth: []
      summary: Update data about company
      tags:
      - Company
  /email/send-code:
    post:
      description: Generates secret code and sends it to specified email
      parameters:
      - description: User's email
        in: body
        name: SignIn
        required: true
        schema:
          $ref: '#/definitions/dto.Email'
      responses:
        "200":
          description: user's session
          schema:
            type: string
        "400":
          description: Data is not valid
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      summary: Send code to specified email
      tags:
      - Auth
  /user:
    patch:
      description: Updates user's info
      parameters:
      - description: Fields to update
        in: body
        name: updFields
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUser'
      responses:
        "200":
          description: Successfully Updated
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errs.MyError'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      security:
      - ApiKeyAuth: []
      summary: Update user's data
      tags:
      - User
  /user/{company_name}:
    get:
      description: Returns PDF file got from user history
      parameters:
      - description: Unique company name from history
        in: path
        name: company_name
        required: true
        type: string
      responses:
        "200":
          description: PDF file
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errs.MyError'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      security:
      - ApiKeyAuth: []
      summary: Generate PDF file from user history
      tags:
      - User
  /user/email:
    patch:
      description: Updates user's email
      parameters:
      - description: New email with password
        in: body
        name: updEmail
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateEmail'
      responses:
        "200":
          description: Successfully Updated
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errs.MyError'
        "401":
          description: User isn't logged in
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      security:
      - ApiKeyAuth: []
      summary: Update user's email
      tags:
      - User
  /user/password:
    patch:
      description: Updates user's password
      parameters:
      - description: Email with new password
        in: body
        name: updPassword
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePassword'
      responses:
        "200":
          description: Successfully Updated
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errs.MyError'
        "404":
          description: User doesn't exist
          schema:
            $ref: '#/definitions/errs.MyError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MyError'
      summary: Update user's password
      tags:
      - User
produces:
- application/json
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: session_id
    type: apiKey
swagger: "2.0"